Move bar chart tile component to tiles directory
