+++ src/lib/ui/tiles/BarChartTile.svelte
